fix: leaderboard table not showing
