+++ src/pages/Leaderboard.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React, { useEffect, useState } from "react";
 import { Card, PageLayout } from "../components";
 import { useLeaderboardStatistics } from "../hooks/stats";
 import { LeaderboardTable } from "../components/Tables";
@@ -6,11 +6,42 @@ import { useAccount } from "wagmi";
 import utils from "../utils";
 import { ethers } from "ethers";
 import { Countdown } from "../components/Countdown";
+import constants from "../constants";
+import dayjs from "dayjs";
+import duration from "dayjs/plugin/duration";
+
+dayjs.extend(duration);
 
 const Leaderboard: React.FC = () => {
   const { stats, balances, userStats, loading } = useLeaderboardStatistics();
   const { isConnected } = useAccount();
-  const [showLeaderboard, setShowLeaderboard] = useState(false);
+
+  const [intervalId, setIntervalId] = useState<NodeJS.Timer>();
+  const [now, setNow] = useState(Date.now());
+
+  const eventTimestamp = +(constants.env.EVENT_TS || "0");
+  const isEventInFuture = dayjs(now).isBefore(dayjs(eventTimestamp));
+
+  useEffect(() => {
+    const interval = setInterval(
+      () => setNow(Date.now()),
+      1 * constants.time.ONE_SECOND_MS
+    );
+    setIntervalId(interval);
+
+    return () => {
+      clearInterval(interval);
+      setIntervalId(undefined);
+    };
+  }, []);
+
+  // clear interval if event is no longer in future
+  useEffect(() => {
+    if (!isEventInFuture && intervalId) {
+      clearInterval(intervalId);
+      setIntervalId(undefined);
+    }
+  }, [isEventInFuture]);
 
   const userHasNoStats = isConnected && !userStats;
 
@@ -26,7 +57,7 @@ const Leaderboard: React.FC = () => {
 
   return (
     <PageLayout>
-      {showLeaderboard ? (
+      {!isEventInFuture ? (
         <React.Fragment>
           {isConnected && (
             <div className="mb-4 flex w-full flex-col justify-center gap-x-1 gap-y-2 text-left md:flex-row lg:justify-between lg:gap-x-4">
@@ -52,9 +83,9 @@ const Leaderboard: React.FC = () => {
         </React.Fragment>
       ) : (
         <Countdown
-          large
-          containerStyles="mt-12"
-          setShowLeaderboard={setShowLeaderboard}
+          eventTimestamp={eventTimestamp}
+          now={now}
+          isEventInFuture={isEventInFuture}
         />
       )}
     </PageLayout>
